Changed ul visibility: hidden to display:none
Changed the ul of the menu from visibility: hidden , to display: none.
Reason being is visibility: hidden hides the ul, but is still there and
that will mess with other extended menus that will be added. Changing it
to display none: removes it from the document , in order to make the
menu function better